+++ lib/recipes/information-systems/recipe.rb
@@ -5,13 +5,18 @@ do |doc, _resources|
   include Kitchen::Directions
 
   book = doc.book
-  _metadata = book.metadata
+  metadata = book.metadata
 
   book.search('cnx-pi').trash
 
   BakePreface.v1(book: book)
   BakeUnnumberedFigure.v1(book: book)
   BakeUnnumberedTables.v1(book: book)
+
+  AddInjectedExerciseId.v1(book: book)
+  book.injected_exercises.each do |exercise|
+    BakeInjectedExercise.v1(exercise: exercise)
+  end
 
   BakeUnitTitle.v1(book: book)
   BakeChapterTitle.v1(book: book)
@@ -34,6 +39,43 @@ do |doc, _resources|
       BakeNumberedTable.v1(table: table,
                            number: "#{chapter.count_in(:book)}.#{table.count_in(:chapter)}")
     end
+
+    # EOC
+    BakeChapterGlossary.v1(chapter: chapter, metadata_source: metadata)
+
+    MoveCustomSectionToEocContainer.v1(
+      chapter: chapter,
+      metadata_source: metadata,
+      container_key: 'summary',
+      uuid_key: '.summary',
+      section_selector: 'section.summary'
+    ) do |section|
+      RemoveSectionTitle.v1(section: section)
+      title = EocSectionTitleLinkSnippet.v1(page: section.ancestor(:page))
+      section.prepend(child: title)
+    end
+
+    eoc_sections = %w[review-questions check-understanding
+                      reflection-questions media-questions]
+
+    eoc_sections.each do |section_key|
+      MoveCustomSectionToEocContainer.v1(
+        chapter: chapter,
+        metadata_source: metadata,
+        container_key: section_key,
+        uuid_key: ".#{section_key}",
+        section_selector: "section.#{section_key}"
+      ) do |section|
+        RemoveSectionTitle.v1(section: section)
+      end
+
+      chapter.composite_pages.search("section.#{section_key}").injected_questions.each do |question|
+        BakeInjectedExerciseQuestion.v1(
+          question: question, number: question.count_in(:composite_page)
+        )
+        BakeFirstElements.v1(within: question)
+      end
+    end
   end
 
   # Appendix
@@ -52,6 +94,7 @@ do |doc, _resources|
 
   BakeTableColumns.v1(book: book)
   BakeFootnotes.v1(book: book)
+  BakeCompositePages.v1(book: book)
   BakeToc.v1(book: book)
   BakeLinkPlaceholders.v1(book: book)
   BakeFolio.v1(book: book)